Vita: in-memory .o2r archives, printf %z fixes, memory tuning
Real-hardware bring-up work.

O2rArchive: read the whole archive into RAM up front via raw sceIoOpen/
sceIoRead/sceIoGetstat and hand libzip a buffer source, instead of letting
it seek through newlib stdio. Real hardware repeatedly faulted inside
sceIoLseek32 via _zip_stdio_op_seek -> _fseeko_r -> __sseek -> _lseek_r,
always at the same archive offset, while the same binary ran fine under
Vita3K. Also avoids repeated microSD seeks. Archives become read-only on
Vita (WriteFile now fails loudly; nothing in the boot path writes).

Also initialise mZipArchive to nullptr - it never was, and the new
early-return paths in Open() let the destructor hand zip_close() a garbage
pointer, which took a data abort.

printf %z: VitaSDK's newlib vsnprintf does not implement the 'z' length
modifier - it emits the literal characters and does NOT consume the
argument, so every remaining argument in the call shifts by one. Where a
%s follows, strlen() then runs on a garbage pointer. Replaced with %u/%x
plus explicit casts (size_t is 32-bit here). ImGui::Text is affected too;
it routes through the same vsnprintf.

vitaGL: disable triple buffering to save a display framebuffer's worth of
VRAM (per Rinnegatamante).

+++ src/fast/backends/gfx_opengl.cpp
@@ -1154,7 +1154,7 @@ static void GLDumpDrawVbo(const float* buf, size_t num_floats, size_t num_tris, 
     if (verts > 6) {
         verts = 6;
     }
-    fprintf(sVboLog, "draw %d tris=%zu:", sGLDumpDrawIndex, num_tris);
+    fprintf(sVboLog, "draw %d tris=%u:", sGLDumpDrawIndex, (unsigned int)num_tris);
     for (size_t i = 0; i < verts; i++) {
         const float* v = buf + i * stride_floats;
         fprintf(sVboLog, "  (%.3f,%.3f,%.3f,%.3f)", v[0], v[1], v[2], v[3]);
